requirements.txt
selenium~=4.8.2
PyYAML~=6.0